Returns the inverse of a matrix from cache if present in cache
These set of functions take a resource intensive operation like
calculating the "Inverse of a matrix" in this case and make it more
efficient.
It does so by checking if the inverse has been calculated previosly and
if so, the function cacheSolve would return the inverted matrix from the
cache instead of re-doing the whole calculation.